pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fashion-clip>=0.2.2",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pinscrape>=3.1.0",
    "scikit-learn>=1.6.1",
    "supabase>=2.13.0",
    "tensorflow>=2.17.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
]
